feat(macros): allow i32/u32/u64/f32/f64 field types
Extend `property_type_for` with arms for the five new numeric Rust
scalars, mapping each to the corresponding `PropertyType` variant added
in the previous commit. Update the unsupported-type error message to
list the new Phase 1 set: `i32`, `i64`, `u32`, `u64`, `f32`, `f64`,
`bool`, `String`.

Rotate the canonical "unsupported field type" UI fixture from `u32` (now
supported) to `u128` (still out of Phase 1) so the trybuild suite keeps
covering the rejection path. The `.stderr` was regenerated via
`TRYBUILD=overwrite cargo test --test trybuild --package frieze` and
only the supported-type list changed; line numbers and span pointers
are stable.

+++ crates/libs/frieze/tests/ui/unsupported_field_type_u128.rs
@@ -2,7 +2,7 @@ use frieze::Schema;
 
 #[derive(Schema)]
 struct S {
-    x: u32,
+    x: u128,
 }
 
 fn main() {}
